first batch of execution span recording.
to get profiler information, invoke gradle with RECORD_SPANS
env variable set to something and find the resulting json
file in build/intermediates/profiler.

so far, only 4 events are generated, this is the first batch.

+++ build-system/builder/src/main/java/com/android/builder/tasks/WorkQueue.java
@@ -87,7 +87,7 @@ public class WorkQueue<T> implements Runnable {
             @NonNull QueueThreadContext<T> queueThreadContext,
             @NonNull String queueName,
             int workforce,
-            int growthTriggerRatio) {
+            float growthTriggerRatio) {
 
         this.mLogger = logger;
         this.mName = queueName;
